changes in codec
